add files and notes aggregates
