+++ web/src/features/responses/components/FormSubmissionDetails/FormSubmissionDetails.tsx
@@ -1,7 +1,7 @@
 import { ArrowTopRightOnSquareIcon } from '@heroicons/react/24/outline';
 import { FlagIcon } from '@heroicons/react/24/solid';
 import { Link, useLoaderData } from '@tanstack/react-router';
-import { format, isEqual } from 'date-fns';
+import { format } from 'date-fns';
 import { Fragment } from 'react';
 import {
   isDateAnswer,
@@ -25,8 +25,6 @@ import { Separator } from '@/components/ui/separator';
 import { Switch } from '@/components/ui/switch';
 import type { FunctionComponent } from '@/common/types';
 import { cn, ratingScaleToNumber } from '@/lib/utils';
-import type { Attachment } from '../../models/form-submission';
-import type { QuestionExtraData } from '../../types';
 import { QuestionExtraDataSection } from '../QuestionExtraDataSection/QuestionExtraDataSection';
 
 export default function FormSubmissionDetails(): FunctionComponent {
@@ -98,32 +96,6 @@ export default function FormSubmissionDetails(): FunctionComponent {
               const answer = formSubmission.answers.find(({ questionId }) => questionId === question.id);
               const notes = formSubmission.notes.filter(({ questionId }) => questionId === question.id);
               const attachments = formSubmission.attachments.filter(({ questionId }) => questionId === question.id);
-              const groupedAttachments = attachments.reduce<Record<string, Attachment[]>>(
-                (grouped, attachment) => ({
-                  ...grouped,
-                  [attachment.timeSubmitted]: [...(grouped[attachment.timeSubmitted] ?? []), attachment],
-                }),
-                {}
-              );
-              const mediaFiles: QuestionExtraData[] = Object.entries(groupedAttachments)
-                .filter(([key]) => !notes.some((note) => isEqual(note.timeSubmitted, key)))
-                .flatMap(([_, attachments]) => ({
-                  questionId: attachments[0]?.questionId ?? '',
-                  text: '',
-                  monitoringObserverId: attachments[0]?.monitoringObserverId ?? '',
-                  timeSubmitted: attachments[0]?.timeSubmitted ?? '',
-                  attachments,
-                }));
-
-              const extraData = [
-                ...notes.map((note) => {
-                  return {
-                    ...note,
-                    attachments: groupedAttachments[note.timeSubmitted] ?? [],
-                  };
-                }),
-                ...mediaFiles,
-              ];
 
               return (
                 <div key={question.id} className='flex flex-col gap-4'>
@@ -190,12 +162,8 @@ export default function FormSubmissionDetails(): FunctionComponent {
                   ) : (
                     '-'
                   )}
-                  {extraData.length > 0 && (
-                    <QuestionExtraDataSection
-                      attachmentCount={attachments.length}
-                      extraData={extraData}
-                      notesCount={notes.length}
-                    />
+                  {(attachments.length > 0 || notes.length > 0) && (
+                    <QuestionExtraDataSection attachments={attachments} notes={notes} />
                   )}
                 </div>
               );
